requirements.txt
faiss-cpu
numpy
openai
sentence-transformers
streamlit
torch
